EDR/sentinelone_files/requirements.txt
pandas
python-dotenv
requests